Battery bug fixes

+++ implementation_and_examples/agent_implementation/agent_control/battery/MicroControllerBatterManager.cpp
@@ -39,10 +39,11 @@ std::pair<float, float> MicroControllerBatteryManager::estimateTransmitConsumpti
     //Check if we have sent a message to this agent recently
 
     int nExchangeIntervalsInPeriod = std::floor( seconds/agent->config.QUADTREE_EXCHANGE_INTERVAL_S); //How many full exchange periods fit into the period
-    double remaining = seconds - nExchangeIntervalsInPeriod * agent->config.QUADTREE_EXCHANGE_INTERVAL_S;
+    float remaining = seconds - nExchangeIntervalsInPeriod * agent->config.QUADTREE_EXCHANGE_INTERVAL_S;
 
     //If the agent has sent a message to this agent recently, we will probably send a message soon
     //Else we will probably not send a message soon
+    //So heuristics based on previous exchanges
 
     int amountOfTransmits = 0;
